requirements.txt
aiogram==3.25.0
python-dotenv==1.2.1
SQLAlchemy==2.0.47
